requirements.txt
psycopg2-binary
faker
python-dotenv
pydantic
pydantic-settings
langchain
langchain-groq
sqlparse
uvicorn
fastapi
langchain-core
langchain-chroma
langchain-huggingface
langchain-text-splitters
langchain-community
chromadb
sentence-transformers
sqlalchemy
sql-metadata
deepeval
langchain-google-genai
langchain-ollama
streamlit
passlib[bcrypt]
pyjwt
python-multipart